Minor content changes and general updates
Updated content on the registration page to mention that you cannot use a Northern Ireland driving licence and on the HMRC IV page to reflect the fact that they dropped the use of payslips and P60.
Updated Gov.uk frontend, DWP and HMRC frontend and node notification client.
Readjusted routes to support new content changes.
Readjusted hmrc iv error page to support new content changes.
Scrapped the Nothern Ireland pages and references as it was never meant to be in the prototype.

GitOrigin-RevId: 45a3711b9e49376cd82983d259333084d6dcf49e